LayoutTests/http/tests/appcache/resource-redirect.html

<html manifest="resources/resource-redirect.manifest">
<body>
<p>Test that a redirect makes resource caching fail.</p>
<p>Should say SUCCESS:</p>
<div id=result></div>
<script>
if (window.layoutTestController) {
    layoutTestController.dumpAsText()
    layoutTestController.waitUntilDone();
}

function log(message)
{
    document.getElementById("result").innerHTML += message + "<br>";
}

function cacheCallback()
{
    log("FAIL: An event was dispatched on window.applicationCache, even though it was not associated with any cache yet.");
}

function timeout()
{
    if (applicationCache.status != applicationCache.UNCACHED) {
        setTimeout(timeout, 100);
        return;
    }

    log("SUCCESS");
    if (window.layoutTestController)
        layoutTestController.notifyDone();
}

applicationCache.addEventListener('cached', function() { log("cached"); cacheCallback() }, false);
applicationCache.addEventListener('noupdate', function() { log("noupdate"); cacheCallback() }, false);
applicationCache.addEventListener('error', function() { log("error"); cacheCallback() }, false);

setTimeout(timeout, 100);

</script>
</body>
</html>
